+++ forui/lib/src/widgets/select/single/select.dart
@@ -59,13 +59,6 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     );
   }
 
-  static Widget _fieldBuilder(BuildContext _, FSelectStyle _, Set<FTextFieldVariant> _, Widget child) => child;
-
-  static Widget _popoverBuilder(BuildContext _, FSelectController<Object?> _, FPopoverController _, Widget content) =>
-      content;
-
-  static String? _validator(Object? _) => null;
-
   /// The control that manages the select's state.
   ///
   /// Defaults to [FSelectControl.managed] if not provided.
@@ -242,7 +235,7 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     FSelectStyleDelta style = const .context(),
     bool autofocus = false,
     FocusNode? focusNode,
-    FFieldBuilder<FSelectStyle> builder = _fieldBuilder,
+    FFieldBuilder<FSelectStyle> builder = FTextField.defaultBuilder,
     FFieldIconBuilder<FTextFieldStyle>? prefixBuilder,
     FFieldIconBuilder<FTextFieldStyle>? suffixBuilder = defaultIconBuilder,
     Widget? label,
@@ -252,7 +245,7 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     VoidCallback? onReset,
     AutovalidateMode autovalidateMode = .onUnfocus,
     String? forceErrorText,
-    FormFieldValidator<T> validator = _validator,
+    FormFieldValidator<T> validator = FFormFieldProperties.defaultValidator,
     Widget Function(BuildContext context, String message) errorBuilder = FFormFieldProperties.defaultErrorBuilder,
     String? hint,
     TextAlign textAlign = .start,
@@ -262,7 +255,7 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     MouseCursor mouseCursor = .defer,
     bool canRequestFocus = true,
     bool clearable = false,
-    FSelectPopoverBuilder<T> popoverBuilder = _popoverBuilder,
+    FSelectPopoverBuilder<T> popoverBuilder = FPopover.defaultPopoverBuilder,
     AlignmentGeometry contentAnchor = .topStart,
     AlignmentGeometry fieldAnchor = .bottomStart,
     FPortalConstraints contentConstraints = const FAutoWidthPortalConstraints(maxHeight: 300),
@@ -410,7 +403,7 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     FSelectStyleDelta style = const .context(),
     bool autofocus = false,
     FocusNode? focusNode,
-    FFieldBuilder<FSelectStyle> builder = _fieldBuilder,
+    FFieldBuilder<FSelectStyle> builder = FTextField.defaultBuilder,
     FFieldIconBuilder<FTextFieldStyle>? prefixBuilder,
     FFieldIconBuilder<FTextFieldStyle>? suffixBuilder = defaultIconBuilder,
     Widget? label,
@@ -420,7 +413,7 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     VoidCallback? onReset,
     AutovalidateMode autovalidateMode = .onUnfocus,
     String? forceErrorText,
-    FormFieldValidator<T> validator = _validator,
+    FormFieldValidator<T> validator = FFormFieldProperties.defaultValidator,
     Widget Function(BuildContext context, String message) errorBuilder = FFormFieldProperties.defaultErrorBuilder,
     String? hint,
     TextAlign textAlign = .start,
@@ -430,7 +423,7 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     MouseCursor mouseCursor = .defer,
     bool canRequestFocus = true,
     bool clearable = false,
-    FSelectPopoverBuilder<T> popoverBuilder = _popoverBuilder,
+    FSelectPopoverBuilder<T> popoverBuilder = FPopover.defaultPopoverBuilder,
     AlignmentGeometry contentAnchor = .topStart,
     AlignmentGeometry fieldAnchor = .bottomStart,
     FPortalConstraints contentConstraints = const FAutoWidthPortalConstraints(maxHeight: 300),
@@ -581,7 +574,7 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     this.style = const .context(),
     this.autofocus = false,
     this.focusNode,
-    this.builder = _fieldBuilder,
+    this.builder = FTextField.defaultBuilder,
     this.prefixBuilder,
     this.suffixBuilder = defaultIconBuilder,
     this.label,
@@ -591,7 +584,7 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     this.onReset,
     this.autovalidateMode = .onUnfocus,
     this.forceErrorText,
-    this.validator = _validator,
+    this.validator = FFormFieldProperties.defaultValidator,
     this.errorBuilder = FFormFieldProperties.defaultErrorBuilder,
     this.hint,
     this.textAlign = .start,
@@ -601,7 +594,7 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     this.mouseCursor = .defer,
     this.canRequestFocus = true,
     this.clearable = false,
-    this.popoverBuilder = _popoverBuilder,
+    this.popoverBuilder = FPopover.defaultPopoverBuilder,
     this.contentAnchor = .topStart,
     this.fieldAnchor = .bottomStart,
     this.contentConstraints = const FAutoWidthPortalConstraints(maxHeight: 300),