+++ forui/lib/src/widgets/slider/slider.dart
@@ -19,9 +19,11 @@ import 'package:forui/src/widgets/tooltip/tooltip_controller.dart';
 /// * [FSliderControl] for configuring a slider's behavior.
 /// * [FSliderStyles] for customizing a slider's appearance.
 class FSlider extends StatelessWidget with FFormFieldProperties<FSliderValue> {
-  static Widget _tooltipBuilder(FTooltipController _, double value) => Text('${(value * 100).toStringAsFixed(0)}%');
+  /// The default tooltip builder that displays the value as a percentage.
+  static Widget defaultTooltipBuilder(FTooltipController _, double value) => Text('${(value * 100).toStringAsFixed(0)}%');
 
-  static String _semanticValueFormatter(double value) => '${(value * 100).toStringAsFixed(0)}%';
+  /// The default semantic value formatter that formats the value as a percentage.
+  static String defaultSemanticValueFormatter(double value) => '${(value * 100).toStringAsFixed(0)}%';
 
   /// The control for the slider.
   final FSliderControl control;
@@ -128,8 +130,8 @@ class FSlider extends StatelessWidget with FFormFieldProperties<FSliderValue> {
     this.trackMainAxisExtent,
     this.trackHitRegionCrossExtent,
     this.tooltipControls = const .new(),
-    this.tooltipBuilder = _tooltipBuilder,
-    this.semanticValueFormatterCallback = _semanticValueFormatter,
+    this.tooltipBuilder = defaultTooltipBuilder,
+    this.semanticValueFormatterCallback = defaultSemanticValueFormatter,
     this.semanticFormatterCallback,
     this.onEnd,
     this.onSaved,